Move global_env_valid to ef_conf_utils
