+++ efopen/ef_conf_utils.py
@@ -20,3 +20,17 @@ def get_template_parameters_file(template_full_path):
       else:
         continue
     return None
+
+def global_env_valid(env):
+  """
+  Given an env, determine if it's a valid "global" or "mgmt" env as listed in EFConfig
+  Args:
+    env: the env to check
+  Returns:
+    True if the env is a valid global env in EFConfig
+  Raises:
+    ValueError with message if the env is not valid
+  """
+  if env not in EFConfig.ACCOUNT_SCOPED_ENVS:
+    raise ValueError("Invalid global env: {}; global envs are: {}".format(env, EFConfig.ACCOUNT_SCOPED_ENVS))
+  return True
